+++ packages/swirl-tokens/dart/lib/styles.light.dart
@@ -4,7 +4,7 @@
 //
 
 // Do not edit directly
-// Generated on Fri, 08 May 2026 09:13:59 GMT
+// Generated on Mon, 18 May 2026 09:56:51 GMT